Tweaked tests
-   Applied some tweaks to the main tests
    -   Grouped some core tests into folders
    -   Some restricted data improvements
    -   Added some null skipper checks
   -   Formatting improvements
-   CHORE: apply similar changes to operations tests
