+++ Packages/Core/tests/Manipulators/ManipulatorExpressionMatchTests.cs
@@ -1,44 +1,43 @@
 namespace ExpressionBuilder.Tests;
 
-[Collection("Operation checks")]
-public class OperationChecks
-{
-    public static IEnumerable<object[]> GetAllOperations() =>
-        AllTests.GetAllOperations();
+using Xunit;
 
+[Collection("Matching queries - Expressions (w/ manipulators)")]
+public class ManipulatorExpressionMatchTests
+{
     [Theory(DisplayName = "Factory")]
-    [MemberData(nameof(GetAllOperations))]
+    [ClassData(typeof(ManipulatorTestData))]
     public void CheckFactoryStrings<T>(TestBuilder<T> match)
         where T : class, IItemable =>
         Assert.Equal(
-            match.Expected.ToMatchString(AllTests.ApplyReplacements),
+            match.Expected.ToMatchString(),
             match.Factory.ToMatchString<T>()
         );
 
     [Theory(DisplayName = "Filterable")]
-    [MemberData(nameof(GetAllOperations))]
+    [ClassData(typeof(ManipulatorTestData))]
     public void CheckFilterableStrings<T>(TestBuilder<T> match)
         where T : class, IItemable =>
         Assert.Equal(
-            match.Expected.ToMatchString(AllTests.ApplyReplacements),
-            match.Filterable.ToMatchString<T>()
+            match.Expected.ToMatchString(),
+            match.Filterable.ToMatchString()
         );
 
     [Theory(DisplayName = "QueryFilterable")]
-    [MemberData(nameof(GetAllOperations))]
+    [ClassData(typeof(ManipulatorTestData))]
     public void CheckQueryableStrings<T>(TestBuilder<T> match)
         where T : class, IItemable =>
         Assert.Equal(
-            match.Expected.ToMatchString(AllTests.ApplyReplacements),
-            match.QueryFilterable.ToMatchString<T>()
+            match.Expected.ToMatchString(ManipulatorTestData.ApplyReplacements),
+            match.QueryFilterable.ToMatchString()
         );
 
     [Theory(DisplayName = "Fluent")]
-    [MemberData(nameof(GetAllOperations))]
+    [ClassData(typeof(ManipulatorTestData))]
     public void CheckFluentStrings<T>(TestBuilder<T> match)
         where T : class, IItemable =>
         Assert.Equal(
-            match.Expected.ToMatchString(AllTests.ApplyReplacements),
-            match.Fluent.ToMatchString<T>()
+            match.Expected.ToMatchString(ManipulatorTestData.ApplyReplacements),
+            match.Fluent.ToMatchString()
         );
 }
