+++ Packages/Core/tests/Operations/OperationTests.cs
@@ -1,47 +1,48 @@
 namespace ExpressionBuilder.Tests;
 
-using System.Collections.Generic;
-using Xunit;
-
-[Collection("Matching queries - Expressions (w/ manipulators)")]
-public class ManipulatorExpressionMatchTests
+[Collection("Operation checks")]
+public class OperationTests
 {
-    public static IEnumerable<object[]> GetAll() =>
-        AllManipulatorTests.GetAllMatchers();
-
     [Theory(DisplayName = "Factory")]
-    [MemberData(nameof(GetAll))]
+    [ClassData(typeof(OperationTestData))]
     public void CheckFactoryStrings<T>(TestBuilder<T> match)
         where T : class, IItemable =>
         Assert.Equal(
-            match.Expected.ToMatchString(),
+            match.Expected.ToMatchString(OperationTestData.ApplyReplacements),
             match.Factory.ToMatchString<T>()
         );
 
     [Theory(DisplayName = "Filterable")]
-    [MemberData(nameof(GetAll))]
+    [ClassData(typeof(OperationTestData))]
     public void CheckFilterableStrings<T>(TestBuilder<T> match)
         where T : class, IItemable =>
         Assert.Equal(
-            match.Expected.ToMatchString(),
+            match.Expected.ToMatchString(OperationTestData.ApplyReplacements),
             match.Filterable.ToMatchString()
         );
 
     [Theory(DisplayName = "QueryFilterable")]
-    [MemberData(nameof(GetAll))]
+    [ClassData(typeof(OperationTestData))]
     public void CheckQueryableStrings<T>(TestBuilder<T> match)
-        where T : class, IItemable =>
+        where T : class, IItemable
+    {
+        var xx = match.Expected.ToMatchString(
+            OperationTestData.ApplyReplacements
+        );
+        var yy = match.QueryFilterable.ToMatchString();
+
         Assert.Equal(
-            match.Expected.ToMatchString(AllManipulatorTests.ApplyReplacements),
+            match.Expected.ToMatchString(OperationTestData.ApplyReplacements),
             match.QueryFilterable.ToMatchString()
         );
+    }
 
     [Theory(DisplayName = "Fluent")]
-    [MemberData(nameof(GetAll))]
+    [ClassData(typeof(OperationTestData))]
     public void CheckFluentStrings<T>(TestBuilder<T> match)
         where T : class, IItemable =>
         Assert.Equal(
-            match.Expected.ToMatchString(AllManipulatorTests.ApplyReplacements),
+            match.Expected.ToMatchString(OperationTestData.ApplyReplacements),
             match.Fluent.ToMatchString()
         );
 }
