+++ Packages/Core/tests/ResultMatchTests.cs
@@ -5,18 +5,12 @@ using ExpressionBuilder.Tests.Models;
 using Xunit;
 
 [Collection("Matching queries - Results")]
-public class ResultAllTests : IClassFixture<ResultMatchSeed>
+public class ResultAllTests(ResultMatchSeed seed)
+    : IClassFixture<ResultMatchSeed>
 {
-    ICollection<Product> _products { get; set; }
-    ICollection<Category> _categories { get; set; }
-    ICollection<StockLocation> _stockLocations { get; set; }
-
-    public ResultAllTests(ResultMatchSeed seed)
-    {
-        _products = seed.Products;
-        _categories = seed.Categories;
-        _stockLocations = seed.StockLocations;
-    }
+    readonly ICollection<Product> _products = seed.Products;
+    readonly ICollection<Category> _categories = seed.Categories;
+    readonly ICollection<StockLocation> _stockLocations = seed.StockLocations;
 
     ICollection<T> GetCollection<T>()
         where T : class =>
@@ -28,10 +22,8 @@ public class ResultAllTests : IClassFixture<ResultMatchSeed>
                     : (ICollection<T>)_categories
             );
 
-    public static IEnumerable<object[]> GetAll() => AllTests.GetAllMatchers();
-
     [Theory(DisplayName = "Factory")]
-    [MemberData(nameof(GetAll))]
+    [ClassData(typeof(CoreTestData))]
     public void CheckFactoryStrings<T>(TestBuilder<T> match)
         where T : class, IItemable
     {
@@ -44,7 +36,7 @@ public class ResultAllTests : IClassFixture<ResultMatchSeed>
     }
 
     [Theory(DisplayName = "Filterable")]
-    [MemberData(nameof(GetAll))]
+    [ClassData(typeof(CoreTestData))]
     public void CheckFilterableStrings<T>(TestBuilder<T> match)
         where T : class, IItemable
     {
@@ -57,7 +49,7 @@ public class ResultAllTests : IClassFixture<ResultMatchSeed>
     }
 
     [Theory(DisplayName = "QueryFilterable")]
-    [MemberData(nameof(GetAll))]
+    [ClassData(typeof(CoreTestData))]
     public void CheckQueryableStrings<T>(TestBuilder<T> match)
         where T : class, IItemable
     {
@@ -70,7 +62,7 @@ public class ResultAllTests : IClassFixture<ResultMatchSeed>
     }
 
     [Theory(DisplayName = "Fluent")]
-    [MemberData(nameof(GetAll))]
+    [ClassData(typeof(CoreTestData))]
     public void CheckFluentStrings<T>(TestBuilder<T> match)
         where T : class, IItemable
     {
