+++ Packages/TestBase/Models/Brand.cs
@@ -1,0 +1,9 @@
+namespace ExpressionBuilder.Tests.Models;
+
+public class Brand
+{
+    public int Id { get; set; }
+
+    public virtual int SupplierId { get; set; }
+    public virtual Supplier Supplier { get; set; } = null!;
+}
