+++ Packages/TestBase/Models/Supplier.cs
@@ -1,0 +1,6 @@
+namespace ExpressionBuilder.Tests.Models;
+
+public class Supplier
+{
+    public int Id { get; set; }
+}
